lots of stuff

+++ Assets/Scripts/Misc/LevelDescription.cs
@@ -14,7 +14,7 @@ public class LevelDescription : ScriptableObject {
     public Dialogue[] dialogue;
 
     [Header("Camera Controls")]
-    public int camSize;
+    public float camSize;
 }
